9518 - WIP working on rewriting the unit tests
