+++ shared/src/business/useCases/trialSessions/generateNoticesForCaseTrialSessionCalendarInteractor.test.js
@@ -1,0 +1,13 @@
+const {
+  applicationContext,
+  testPdfDoc,
+} = require('../../test/createTestApplicationContext');
+const {
+  generateNoticesForCaseTrialSessionCalendarInteractor,
+} = require('./generateNoticesForCaseTrialSessionCalendarInteractor');
+
+describe('generateNoticesForCaseTrialSessionCalendarInteractor', () => {
+  beforeAll(() => {});
+
+  it('does stuff', async () => {});
+});
